v1.1.0
Added a link to a live demo of the current release candidate to the project's readme file.

Added publish date to individual posts in archive, category and tag pages.

Improved the dark mode scheme with more complimentary colors.

Removed references to the now defunct Noir project

Updated Ruby gem to 1.1.0